moved strict mapping tests and added one more test
