+++ test/src/kbasesearchengine/test/search/ElasticIndexingStorageTest.java
@@ -4,6 +4,10 @@ import static kbasesearchengine.test.common.TestCommon.set;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 import java.io.File;
 import java.io.IOException;
@@ -23,6 +27,11 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
+import ch.qos.logback.classic.Level;
+import ch.qos.logback.classic.Logger;
+import ch.qos.logback.classic.spi.LoggingEvent;
+import ch.qos.logback.core.Appender;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableList;
 import org.apache.commons.io.FileUtils;
 import org.apache.http.HttpHost;
@@ -71,6 +80,8 @@ import kbasesearchengine.test.common.TestCommon;
 import kbasesearchengine.test.controllers.elasticsearch.ElasticSearchController;
 import kbasesearchengine.test.parse.SubObjectExtractorTest;
 import org.junit.rules.ExpectedException;
+import org.mockito.ArgumentMatcher;
+import org.slf4j.LoggerFactory;
 import us.kbase.common.service.UObject;
 
 public class ElasticIndexingStorageTest {
@@ -1536,5 +1547,102 @@ public class ElasticIndexingStorageTest {
         assertThat("no data stored", res.guids.isEmpty(), is(true));
         assertThat("no data stored", res.objects, is((Set<ObjectData>) null));
         assertThat("no data stored", res.total, is(0));
+    }
+
+    @Test
+    public void strictMappingFail() throws Exception {
+        ((ch.qos.logback.classic.Logger) org.slf4j.LoggerFactory
+                .getLogger(org.slf4j.Logger.ROOT_LOGGER_NAME))
+                .setLevel(Level.ALL);
+
+        Logger root = (Logger) LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
+        final Appender mockAppender = mock(Appender.class);
+        when(mockAppender.getName()).thenReturn("MOCK");
+        root.addAppender(mockAppender);
+
+        final Map<String, Object> data = new HashMap<>();
+        data.put("source", "a long string");
+        data.put("title", "a title");
+        data.put("newField", "new field value");  // that does not exist in map
+
+
+        indexStorage.indexObjects(
+                ObjectTypeParsingRules.getBuilder(
+                        new SearchObjectType("Narrative", 1),
+                        new StorageObjectType("WS", "Narrative"))
+                        .withIndexingRule(IndexingRules.fromPath(new ObjectJsonPath("source"))
+                                .build())
+                        .withIndexingRule(IndexingRules.fromPath(new ObjectJsonPath("title"))
+                                .build())
+                        .build(),
+                SourceData.getBuilder(new UObject(new HashMap<>()), "objname1", "creator1")
+                        .build(),
+                Instant.ofEpochMilli(10000),
+                null,
+                new GUID("WS:1/1/1"),
+                ImmutableMap.of(new GUID("WS:1/1/1"), new ParsedObject(
+                        new ObjectMapper().writeValueAsString(data),
+                        data.entrySet().stream().collect(Collectors.toMap(
+                                e -> e.getKey(), e -> Arrays.asList(e.getValue()))))),
+                false);
+
+        // verify error log was written
+        verify(mockAppender).doAppend(argThat(new ArgumentMatcher() {
+            @Override
+            public boolean matches(final Object argument) {
+                String msg = ((LoggingEvent)argument).getFormattedMessage();
+                return msg.contains("status=400") &&
+                        msg.contains("type=strict_dynamic_mapping_exception") &&
+                        msg.contains("reason=mapping set to strict, dynamic introduction of " +
+                                "[newField] within [key] is not allowed");
+            }
+        }));
+    }
+
+
+    @Test
+    public void strictMappingPass() throws Exception {
+        ((ch.qos.logback.classic.Logger) org.slf4j.LoggerFactory
+                .getLogger(org.slf4j.Logger.ROOT_LOGGER_NAME))
+                .setLevel(Level.ALL);
+
+        // input data fits mapping
+        final Map<String, Object> data = new HashMap<>();
+        data.put("source", "a long string");
+        data.put("title", "a title");
+
+        Logger root = (Logger) LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
+        final Appender mockAppender = mock(Appender.class);
+        when(mockAppender.getName()).thenReturn("MOCK");
+        root.addAppender(mockAppender);
+
+        indexStorage.indexObjects(
+                ObjectTypeParsingRules.getBuilder(
+                        new SearchObjectType("Narrative", 1),
+                        new StorageObjectType("WS", "Narrative"))
+                        .withIndexingRule(IndexingRules.fromPath(new ObjectJsonPath("source"))
+                                .build())
+                        .withIndexingRule(IndexingRules.fromPath(new ObjectJsonPath("title"))
+                                .build())
+                        .build(),
+                SourceData.getBuilder(new UObject(new HashMap<>()), "objname1", "creator1")
+                        .build(),
+                Instant.ofEpochMilli(10000),
+                null,
+                new GUID("WS:1/1/1"),
+                ImmutableMap.of(new GUID("WS:1/1/1"), new ParsedObject(
+                        new ObjectMapper().writeValueAsString(data),
+                        data.entrySet().stream().collect(Collectors.toMap(
+                                e -> e.getKey(), e -> Arrays.asList(e.getValue()))))),
+                false);
+
+        // verify info log was written
+        verify(mockAppender).doAppend(argThat(new ArgumentMatcher() {
+            @Override
+            public boolean matches(final Object argument) {
+                String msg = ((LoggingEvent)argument).getFormattedMessage();
+                return msg.contains("Indexed 1 item(s).");
+            }
+        }));
     }
 }
